refactored cmd line options; added plain key-value output
