+++ scripts/processPerfServletData.py
@@ -5,25 +5,25 @@ This script processes an output file as generated by the performance servlet of 
 @author <a href="hhuebler@2innovate.at">Hermann Huebler</a>
 '''
 #######################
+from types import DictType, ListType, IntType
+from utils import options as o
+from utils import logger as l
 import sys
 reload(sys)
-import getopt
 import os
 import os.path
 import re
 import ssl
-from utils import logger as l
 ##### import xml.dom
 ##### import xml.dom.minidom
 import datetime
 import time
-import codecs
 import json
 import xml.etree.ElementTree as ET
 import urllib
 import urllib2
 import httplib
-from types import DictType, ListType, StringType, UnicodeType, IntType
+
 # Default encoding should be utf-8
 sys.setdefaultencoding('utf8')
 # emulate Boolean
@@ -43,30 +43,6 @@ def wcvGetTime(sep=""):
     Returns the time as hhmmss
     '''
     return str(time.strftime("%H" + sep + "%M" + sep + "%S"))
-
-
-@l.logEntryExit
-def printUsage(scriptName):
-    '''
-    '''
-    print "Usage:", scriptName, "--help (this information) | [--xml|-x <perfServletXmlFile> | (--url |-u <perfServletUrl> | --seconds|-s <seconds> --wasUser <was_user> --wasPassword <wasPassword>]) [--json|-j <json_outfile>] [--cell|-c <was_cell_name>] [--noempty|-n] [--replace|-r] [--influxUrl|-i <url> --influxDb|-d <dbName>] [--omitSummary|-o [-U|--targetUser <user> -P|--targetPwd <password>]]"
-    print "\tWhereby:"
-    print "\t\t<perfServletXmlFile>\t\tname of the file with the WAS performance servlet output. Mutual exclusise with <perfServletXmlFile>"
-    print "\t\t<perfServletUrl>\t\tfull URL of the performance servlet to get the data. For example:"
-    print "\t\t\t\t\t\thttp://<host>:<port>/wasPerfTool/servlet/perfservlet?node=<nodeName>&server=<serverName>&module=connectionPoolModule+jvmRuntimeModule"
-    print "\t\t<seconds>\t\t\tinterval in seconds between fetching <perfServletUrl>"
-    print "\t\t<json_outfile>\t\t\tname of the file to which the JSON output will be written"
-    print "optional:"
-    print "\t\t<was_cell_name>\t\t\tname of the WAS cell being used as the root of the tags. Defaults to: \"cell\""
-    print "\t\t<--noempty|-n>\t\t\tremove empty metrics. Defaults to false"
-    print "\t\t<--replace|-r>\t\t\treplace the <json_outfile> if the file exists. Defaults to false"
-    print "\t\t<--omitSummary|-o>\t\tomit summary measurement like for example measurement for JDBC providers"
-    print "\t\t<--influxUrl|-i>\t\trest URL for InfluxDb to which data should be posted"
-    print "\t\t<--influxDb|-d>\t\t\tinfluxDb database to which data should be posted. You might attach a retention policy like: \n\t\t\t\t\t\t\"CREATE RETENTION POLICY TWO_WEEKS ON pmidata DURATION 2w REPLICATION 1\""
-    print "\t\t<--targetUser|-U>\t\tuser name to authenticate on the target plattform (for example influxDb)"
-    print "\t\t<--targetPwd|-d>\t\tpassword being used to authenticate on the target plattform (for example influxDb)"
-    print "\t\t<--wasUser>\t\t\tuser name to authenticate against WebSphere to retrieve the performance servlet data"
-    print "\t\t<--wasPassword>\t\t\tpassword o authenticate against WebSphere to retrieve the performance servlet data"
 
 
 @l.logEntryExit
@@ -217,154 +193,6 @@ def getFieldsString(perfDataDictList):
 
     l.debug("Returning fields string: \n'%s'" % (rtnString))
     return rtnString
-
-
-@l.logEntryExit
-def parmDictToTuple(parmDict):
-    '''
-    Parameter dictonary to variable tuple
-    '''
-    ##
-    ## Key values for the parameters dictonary
-    parmKeyList = ["help", "xml", "json", "cell", "noempty", "replace", "influxUrl", "influxDb", "omitSummary", "url", "seconds", "user", "password", "wasUser", "wasPassword"]
-    ##
-    ## Copy the parameter values to the variables
-    for key in parmKeyList:
-        if (key == "xml"):
-            if (parmDict.get(key) != None):
-                parmServletXmlFile = parmDict[key]
-            else:
-                parmServletXmlFile = None
-            l.debug("XMLFile=%s" % (parmServletXmlFile))
-        elif (key == "url"):
-            if (parmDict.get(key) != None):
-                parmPerfServletUrl = parmDict[key]
-            else:
-                parmPerfServletUrl = None
-            l.debug("PerfServletUrl=%s" % (parmPerfServletUrl))
-        elif (key == "json"):
-            if (parmDict.get(key) != None):
-                parmJsonOutFileName = parmDict[key]
-            else:
-                parmJsonOutFileName = None
-            l.debug("jsonOutFile=%s" % (parmJsonOutFileName))
-        elif (key == "cell"):
-            parmWasCellName = parmDict[key]
-            l.debug("CellName=%s" % (parmWasCellName))
-        elif (key == "noempty"):
-            parmNoEmpty = parmDict[key]
-            l.debug("noempty=%s" % (parmNoEmpty))
-        elif (key == "replace"):
-            parmReplace = parmDict[key]
-            l.debug("replace=%s" % (parmReplace))
-        elif (key == "omitSummary"):
-            parmOmitSummary = parmDict[key]
-            l.debug("omitSummary=%s" % (parmOmitSummary))
-        elif (key == "influxUrl"):
-            if (parmDict.get(key) != None):
-                parmInfluxUrl = parmDict[key]
-            else:
-                parmInfluxUrl = None
-            l.debug("influxUrl=%s" % (parmInfluxUrl))
-        elif (key == "influxDb"):
-            if (parmDict.get(key) != None):
-                parmInfluxDb = parmDict[key]
-            else:
-                parmInfluxDb = None
-            l.debug("influxDb=%s" % (parmInfluxDb))
-        elif (key == "seconds"):
-            if (parmDict.get(key) != None):
-                parmSeconds = parmDict[key]
-            else:
-                parmSeconds = None
-            l.debug("parmSeconds=%s" % (parmSeconds))
-        elif (key == "user"):
-            if (parmDict.get(key) != None):
-                parmTargetUser = parmDict[key]
-            else:
-                parmTargetUser = None
-            l.debug("parmTargetUser=%s" % (parmTargetUser))
-        elif (key == "password"):
-            if (parmDict.get(key) != None):
-                parmTargetPwd = parmDict[key]
-            else:
-                parmTargetPwd = None
-            l.debug("parmTargetPwd=%s" % (parmTargetPwd))
-        elif (key == "wasUser"):
-            if (parmDict.get(key) != None):
-                parmWasUser = parmDict[key]
-            else:
-                parmWasUser = None
-            l.debug("parmWasUser=%s" % (parmWasUser))
-        elif (key == "wasPassword"):
-            if (parmDict.get(key) != None):
-                parmWasPwd = parmDict[key]
-            else:
-                parmWasPwd = None
-            l.debug("parmWasPwd=%s" % (parmWasPwd))
-        else:
-            pass
-    ##
-    ## Return the variables
-    return (parmServletXmlFile, parmPerfServletUrl, parmJsonOutFileName, parmWasCellName, parmNoEmpty, parmReplace, parmOmitSummary, parmInfluxUrl, parmInfluxDb, parmSeconds, parmTargetUser, parmTargetPwd, parmWasUser, parmWasPwd)
-
-
-@l.logEntryExit
-def getParmDict(opts, scriptName):
-    '''
-    Returns a dictionary of the parameters. Key is the long option without '-'s and data is the value
-    '''
-    if (len(opts) == 0):
-        printUsage (scriptName)
-        sys.exit(-1)
-
-    rtnDict = {}
-    l.debug("Setting default values for optional parameters")
-    rtnDict["noempty"] = False
-    rtnDict["replace"] = False
-    rtnDict["omitSummary"] = False
-    rtnDict["cell"] = "cell"
-    rtnDict["seconds"] = 0
-
-    l.debug("Copying CLI options to parameter dictionary ...")
-    for option, value in opts:
-        l.debug("option: '%s'; value: '%s'" % (option, value))
-        if option in ("-?", "--help"):
-            rtnDict["help"] = value
-        elif option in ("-x", "--xml"):
-            rtnDict["xml"] = value
-        elif option in ("-j", "--json"):
-            rtnDict["json"] = value
-        elif option in ("-c", "--cell"):
-            rtnDict["cell"] = value
-        elif option in ("-n", "--noempty"):
-            rtnDict["noempty"] = True
-        elif option in ("-r", "--replace"):
-            rtnDict["replace"] = True
-        elif option in ("-i", "--influxUrl"):
-            rtnDict["influxUrl"] = value
-        elif option in ("-d", "--influxDb"):
-            rtnDict["influxDb"] = value
-        elif option in ("-o", "--omitSummary"):
-            rtnDict["omitSummary"] = True
-        elif option in ("-u", "--url"):
-            rtnDict["url"] = value
-        elif option in ("-s", "--seconds"):
-            rtnDict["seconds"] = int(value)
-        elif option in ("-U", "--targetUser"):
-            rtnDict["user"] = value
-        elif option in ("-P", "--targetPwd"):
-            rtnDict["password"] = value
-        elif option in ("--wasUser"):
-            rtnDict["wasUser"] = value
-        elif option in ("--wasPassword"):
-            rtnDict["wasPassword"] = value
-
-        else:
-            l.error("Unsupported option '%s' provided. Exiting ..." % (option))
-            sys.exit(1)
-
-    return rtnDict
 
 
 @l.logEntryExit
@@ -519,208 +347,6 @@ def getPerfServletData(perfServletUrl, wasUser=None, wasPassword=None):
 
 
 @l.logEntryExit
-def checkParm(parmDict, scriptName):
-    '''
-    Checks and verifies the provided parameters
-    '''
-    l.debug("Start checking parameters")
-    xml = jsonFile = influxUrl = influxDb = url = targetUser = targetPwd = wasUser = wasPassword = ""
-
-    for parmKey in parmDict.keys():
-        l.debug("parmKey: '%s'; parmValue: '%s'" % (parmKey, parmDict[parmKey]))
-        if (parmKey == "help"):
-            printUsage (scriptName)
-            sys.exit(0)
-        elif (parmKey == "xml"):
-            if (parmDict[parmKey] == None):
-                raise Exception, "Parameter -x|--xml requires a value ..."
-            else:
-                xml = parmDict[parmKey]
-        elif (parmKey == "json"):
-            if (parmDict[parmKey] == None):
-                raise Exception, "Parameter -j|--json requires a value ..."
-            else:
-                jsonFile = parmDict[parmKey]
-        elif (parmKey == "cell"):
-            l.debug("cell name provided is: '%s'" % (parmDict[parmKey]))
-        elif (parmKey == "noempty"):
-            if (parmDict[parmKey] == True):
-                l.debug("Empty values will be removed ...")
-        elif (parmKey == "replace"):
-            if (parmDict[parmKey] == True):
-                l.debug("Existing JSON file will be replaced if it exists ...")
-            replaceJson = parmDict[parmKey]
-        elif (parmKey == "influxUrl"):
-            if (parmDict[parmKey] == None):
-                raise Exception, "Parameter -i|--influxUrl requires a value ..."
-            else:
-                influxUrl = parmDict[parmKey]
-        elif (parmKey == "influxDb"):
-            if (parmDict[parmKey] == None):
-                raise Exception, "Parameter -d|--influxDb requires a value ..."
-            else:
-                influxDb = parmDict[parmKey]
-        elif (parmKey == "url"):
-            if (parmDict[parmKey] == None):
-                raise Exception, "Parameter -u|--url requires a value ..."
-            else:
-                url = parmDict[parmKey]
-        elif (parmKey == "omitSummary"):
-            if (parmDict[parmKey] == True):
-                l.debug("Summary statistics will be omitted ...")
-        elif (parmKey == "seconds"):
-            if (parmDict[parmKey] == None):
-                raise Exception, "Parameter -s|--seconds requires a value ..."
-            else:
-                seconds = parmDict[parmKey]
-        elif (parmKey == "user"):
-            if (parmDict[parmKey] == None):
-                raise Exception, "Parameter -U|--targetUser requires a value ..."
-            else:
-                targetUser = parmDict[parmKey]
-        elif (parmKey == "password"):
-            if (parmDict[parmKey] == None):
-                raise Exception, "Parameter -P|--targetPwd requires a value ..."
-            else:
-                targetPwd = parmDict[parmKey]
-        elif (parmKey == "wasUser"):
-            if (parmDict[parmKey] == None):
-                raise Exception, "Parameter --wasUser requires a value ..."
-            else:
-                wasUser = parmDict[parmKey]
-        elif (parmKey == "wasPassword"):
-            if (parmDict[parmKey] == None):
-                raise Exception, "Parameter --wasPassword requires a value ..."
-            else:
-                wasPassword = parmDict[parmKey]
-        else:
-            l.error("Unsupported option '%s' provided. Exiting ..." % (option))
-            sys.exit(1)
-    ##
-    ## If we have an xml input file it must exist and be readable
-    if (xml != ""):
-        ##
-        ## XML file must exist. i.e. we need to be able to read the xml file!
-        if (not os.path.isfile(xml)):
-            raise Exception, 'XML file specified does not exist'
-        ##
-        ## Can we open the XML file for read?
-        try:
-            inFile = open(xml, "rU")
-            inFile.close()
-        except IOError:
-            raise Exception, 'XML file can not be read. Check file permissions'
-    if (jsonFile != ""):
-        ##
-        ## JSON file must not exist unless --replace is specified.
-        if (os.path.isfile(jsonFile)):
-            if (replaceJson != True):
-                raise Exception, 'JSON Existis and --replace was not specified'
-        ##
-        ## Can we open the JSON file for write?
-        try:
-            ##
-            ## We build a temporary file name to test if we can create a file at the location provided
-            tmpFileName = jsonFile + ".tmp"
-            inFile = open(tmpFileName, "w")
-            ##
-            ## Cleanup the temporary file
-            inFile.close()
-            os.remove(tmpFileName)
-        except IOError:
-            raise Exception, 'JSON file can not be created. Check file permissions'
-    ##
-    ## --url and --xml are mutual exclusive
-    if (((url != "") and (xml != "")) or ((url == "") and (xml == ""))):
-        raise Exception, "--xml and --url are mutual exclusive but one is required"
-    ##
-    ## If we get an url we need an interval in seconds as well
-    if ((url != "") and (parmDict["seconds"] == None)):
-        raise Exception, "Specifying an url requires to specify an interval in seconds as well"
-    ##
-    ## If an interval in seconds is provided it must be an integer
-    if (parmDict.get("seconds") != None):
-        if (not isinstance(parmDict["seconds"], IntType)):
-            raise Exception, 'seconds must be an integer'
-    ##
-    ## If influxUrl is specificied we need also influxDb and vice versa
-    if ((influxUrl != "") and ((influxDb == None) or (influxDb == ""))) or ((influxDb != "") and ((influxUrl == None) or (influxUrl == ""))):
-        raise Exception, 'influxUrl and influxDb must be specified together'
-    ##
-    ## If we have a user we need a password and vice versa
-    if ((targetUser != "") and ((targetPwd == None) or (targetPwd == ""))) or ((targetUser != "") and ((targetUser == None) or (targetUser == ""))):
-        raise Exception, 'targetUser and targetPwd must be specified together'
-    ##
-    ## If we have a wasUser we need a wasPassword as well and vice versa
-    if ((wasUser != "") and ((wasPassword == None) or (wasPassword == ""))) or ((wasPassword != "") and ((targetUser == None) or (targetUser == ""))):
-        raise Exception, 'wasUser and wasPassword must be specified together'
-    ##
-    ## Validata influxDb URL if provided
-    if (influxUrl != ""):
-        ##
-        ## If we have an influxUrl let's check if we can get a ping request
-        try:
-            (urlSchema, urlHost, urlPort) = splitHttpUrlString(influxUrl)
-        except Exception as e:
-            raise Exception, sys.exc_info()[1]
-
-        httpConn = getHttpConnection(urlSchema, urlHost, urlPort)
-        try:
-            tmpUri = "/ping"
-            tmpUri += buildQueryString(u=targetUser, p=targetPwd)
-            l.debug("Uri to /ping Influx: '%s'" % (tmpUri))
-            httpConn.request("GET", tmpUri)
-            httpResponse = httpConn.getresponse()
-            httpConn.close()
-        except Exception as e2:
-            errorString = "Failed to get ping from influx, '%s'" % (e2.strerror)
-            raise Exception, errorString
-        ##
-        ## indluxDb ping returns code 204
-        if (httpResponse.status != httplib.NO_CONTENT):
-            httpConn.close()
-            ##
-            ## If authentication is required
-            if (httpResponse.status == httplib.UNAUTHORIZED):
-                raise Exception, 'influx database requires authentication'
-            else:
-                raise Exception, 'influxUrl is incorrect or influx db is not running'
-        else:
-            l.debug("influx URL ping returned status code: '%d'", httpResponse.status)
-        ##
-        ## Check if the database exists
-        tmpUri = "/query"
-        tmpUri += buildQueryString(u=targetUser, p=targetPwd, q="show databases")
-        l.debug("Uri to /query to check for databases: '%s'" % (tmpUri))
-        httpConn.request("GET", tmpUri)
-        httpResponse = httpConn.getresponse()
-        ##
-        ## Get the response json
-        responseData = httpResponse.read()
-        l.debug("influx DB query databases: '%s'" % (str(responseData)))
-        httpConn.close()
-        influxDebugVarsDict  = json.loads(responseData)
-        ##
-        ## {"results":[{"statement_id":0,"series":[{"name":"databases","columns":["name"],"values":[["pmidata"],["_internal"],["hhuedb"]]}]}]}
-        influxSeriesResultList = influxDebugVarsDict["results"]
-        for influxSeriesResult in influxSeriesResultList:
-            if (influxSeriesResult.get("series") != None):
-                influxSeriesDictList = influxSeriesResult["series"]
-                l.debug("influxSeriesDictList: '%s'" % (str(influxSeriesDictList)))
-                for influxSeriesDict in influxSeriesDictList:
-                    l.debug("influxSeriesDict: '%s'" % (str(influxSeriesDict)))
-                    if (influxSeriesDict.get("name") == "databases"):
-                        databaseSeriesDbList = influxSeriesDict["values"]
-                        l.debug("databaseSeriesDbList: '%s'" % (str(databaseSeriesDbList)))
-                        validDb = False
-                        for databaseSeriesDb in databaseSeriesDbList:
-                            if (influxDb in databaseSeriesDb):
-                                validDb = True
-                        if (validDb != True):
-                            raise Exception, 'influxDb not found at influxUrl'
-
-
-@l.logEntryExit
 def buildDictFromXmlNode(statistics_node):
     '''
     Converts the statistics_node to a dictionary with the attributes in List attrList
@@ -755,19 +381,11 @@ def processLeafNode(xmlNode, parentNodeNames):
         l.debug("Checking for statistic classification: '%s'" % (statClassification))
         xPathString = "./" + statClassification
         for statistics_node in xmlNode.findall(xPathString):
-            l.debug("Processing node: '%s'" % (statistics_node.get("name")))
+            l.debug("Processing node: '%s'" % (statistics_node.get("name", "N/A")))
             statisticDataDict = {}
             ##
             ##  Depending on the node tag we have to build out dictionary
-            if (statistics_node.tag == "CountStatistic"):
-                statisticDataDict = buildDictFromXmlNode(statistics_node)
-            elif (statistics_node.tag == "AverageStatistic"):
-                statisticDataDict = buildDictFromXmlNode(statistics_node)
-            elif (statistics_node.tag == "TimeStatistic"):
-                statisticDataDict = buildDictFromXmlNode(statistics_node)
-            elif (statistics_node.tag == "RangeStatistic"):
-                statisticDataDict = buildDictFromXmlNode(statistics_node)
-            elif (statistics_node.tag == "BoundedRangeStatistic"):
+            if statistics_node.tag in STATISTIC_CLASSIFICATION_LIST:
                 statisticDataDict = buildDictFromXmlNode(statistics_node)
             else:
                 l.error("Invalid tag of XML node found: '%s'. Exiting ..." % (statistics_node.tag))
@@ -832,7 +450,7 @@ def getStatsData(parentNodeNames, xmlNode):
 
     else:
         ##
-        ## No sub Stat nodes --> a leave found.
+        ## No sub Stat nodes --> a leaf found.
         l.debug("Leaf node: '%s' found" % (xmlNode.get("name")))
         subStatNodeList = []
         leafNodeResult = processLeafNode(xmlNode, parentNodeNames)
@@ -862,11 +480,11 @@ def getNodes(wasCellName, root):
         if (responseStatus != 'success'):
             l.error("PerformanceMonitor responseStatus indicates an invalid file!")
         else:
-            l.info("PerformanceMonitor responseStatus indicates a valid file!")
+            l.verbose("PerformanceMonitor responseStatus indicates a valid file!")
     ##
     ## process sub-nodes
     for nodeNode in root.findall("./Node"):
-        l.info("Processing xmlNode.tag: '%s'" % (nodeNode.tag))
+        l.verbose("Processing xmlNode.tag: '%s'" % (nodeNode.tag))
         parentNodeNames += wasCellName
         ##
         ## Node nodes
@@ -998,6 +616,55 @@ def writeToInflux(parmInfluxUrl, parmInfluxDb, parmTargetUser, parmTargetPwd, pe
     l.info("writeToInflux: Number of rows inserted: '%d'" % (rowCount))
 
 
+@l.logEntryExit
+def outputFormatter(perfList, outFormat=None):
+    '''
+    formats data as specified by outFormat and returns a String
+    '''
+    ##
+    ## All the values are reported by the current time in ms
+    unixTime = str(time.time()).replace(".", "") + "0"
+    ##
+    ## influxdb write end-point with query string
+
+    ##
+    ## Number of rows inserted
+    rowCount = 0
+    postDataList = []
+    ##
+    ## iterate over the perflist and build the REST API string.
+    ## The "tags" is string of tags separated by NODE_SEPARATOR and the counters will be the fields
+    for perfListEntry in perfList:
+        ##
+        ## Get the tags string
+        xmlTagString = perfListEntry["tags"]
+        perfDataDictList = perfListEntry["perfdata"]
+        # l.info(perfListEntry)
+        ##
+        ## get the measure from the tags
+        measurement = getMeasurement(xmlTagString)
+        tagsString = getTagsString(xmlTagString)
+        fieldsString = getFieldsString(perfDataDictList)
+        tagList = tagsString.split(",")
+        fieldList = fieldsString.split(",")
+        if l.isDebugEnabled():
+            l.debug("Tag list:")
+            for f in tagList:
+                l.debug("-  %s", f)
+            l.debug("field list:")
+            for f in fieldList:
+                l.debug("-  %s", f)
+
+        fieldsString = " ".join(fieldList)
+        tagsString = " ".join(tagList)
+        postData = "{}-{} {} {}".format(wcvGetDate(), wcvGetTime(), tagsString, fieldsString)
+        print postData
+        postDataList.append(postData)
+
+    l.verbose("Number of rows returned: '%d'" % (len(postDataList)))
+    return "\n".join(postDataList)
+
+
 '''
 Main ... the start of the execution :)
 '''
@@ -1010,30 +677,14 @@ def main():
     l.info("Script: %s/%s" % (WCV_SCRIPT_PATH, WCV_SCRIPTNAME))
     l.info("Date = %s / Time = %s" % (wcvGetDate("/"), wcvGetTime(":")))
     l.info("%s" % (SEPARATOR_LINE))
+    
     ##
     ## parse CLI command line arguments
     l.debug("sys.argv=%s" % (sys.argv))
     sysArgv = sys.argv[1:]
-    try:
-        opts, args = getopt.getopt(sysArgv, "hx:j:c:nri:d:u:os:U:P:", ["help", "xml=", "json=", "cell=", "noempty", "replace", "influxUrl=", "influxDb=", "url=", "omitSummary", "seconds=", "targetUser=", "targetPwd=", "wasUser=", "wasPassword="])
-        parmArgs = args[1:]
-        l.debug("Got the following paramters: opts:'%s'; args: '%s'" % (str(opts), str(args)))
-    except getopt.GetoptError, err:
-        printUsage(WCV_SCRIPTNAME)
-        l.error(">> Called exception: '%s'" %(str(err)))
-        sys.exit(2)
-    ##
-    ## Check the parameters being passed
-    try:
-        parmDict = getParmDict(opts, WCV_SCRIPTNAME)
-        checkParm(parmDict, WCV_SCRIPTNAME)
-    except (Exception), err:
-        printUsage(WCV_SCRIPTNAME)
-        print ">> Caught exception: " + str(err) + " by checkParm"
-        return 2
-    ##
     ## Copy the parameter values to the variables
-    (parmServletXmlFile, parmPerfServletUrl, parmJsonOutFileName, parmWasCellName, parmNoEmpty, parmReplace, parmOmitSummary, parmInfluxUrl, parmInfluxDb, parmSeconds, parmTargetUser, parmTargetPwd, parmWasUser, parmWasPwd) = parmDictToTuple(parmDict)
+    (parmServletXmlFile, parmPerfServletUrl, parmJsonOutFileName, parmWasCellName, parmNoEmpty, parmReplace, parmOmitSummary, parmInfluxUrl,
+     parmInfluxDb, parmSeconds, parmTargetUser, parmTargetPwd, parmWasUser, parmWasPwd, parmOutFileName, parmOutFormat) = o.parseArguments(WCV_SCRIPTNAME, sysArgv)
 
     curDate = str(datetime.datetime.now().date())
     curTime = datetime.datetime.now().strftime("%H:%M:%S")
@@ -1081,7 +732,8 @@ def main():
                 perfList = removeSummaryData(perfList)
             ##
             ## Write data to the outfile if selected
-            ## l.debug("FINALLY: '%s'" % (str(json.dumps(perfList))))
+            l.debug("FINALLY: '%s'" % (str(json.dumps(perfList))))
+            
             if (parmJsonOutFileName != None):
                 outFile = open(parmJsonOutFileName, "w")
                 outFile.write(str(json.dumps(perfList)))
@@ -1093,6 +745,14 @@ def main():
             if ((parmInfluxUrl != None) and (parmInfluxUrl != "")):
                 writeToInflux(parmInfluxUrl, parmInfluxDb, parmTargetUser, parmTargetPwd, perfList)
                 l.info("Data pushed to influx DB")
+
+            ##
+            ## append to log file in Splunk-like format "TS key1=value1 key2=value2 ..."
+            if (parmOutFileName != None):
+                    outFile = open(parmOutFileName, "w")
+                    outFile.write(outputFormatter(perfList, parmOutFormat))
+                    outFile.close()
+                    
         ##
         ## If input was a file we break the loop otherwise we sleep
         if (parmServletXmlFile != None):
@@ -1116,7 +776,7 @@ if (os.environ.get("TWOI_LOG_LEVEL") == None):
     l.setLoglevel(l.LogLevels.INFO)
 else:
     l.info("Environment variable 'TWOI_LOG_LEVEL' found. Setting Loglevel to: '%s'" % os.environ.get("TWOI_LOG_LEVEL"))
-    l.setLoglevelByName(os.environ.get("TWOI_LOG_LEVEL"))
+    l.setLoglevelByName(os.environ.get("TWOI_LOG_LEVEL").upper())
 l.debug("Created default logger ...")
 ##
 ## Disable SSL Certificate verification
